+++ desktop/src/main/java/com/onecat/burst/ui/displays/RegularEmitterDisplay.java
@@ -2,20 +2,22 @@ package com.onecat.burst.ui.displays;
 
 import com.badlogic.gdx.graphics.g3d.particles.emitters.RegularEmitter;
 import com.badlogic.gdx.scenes.scene2d.Actor;
-import com.badlogic.gdx.scenes.scene2d.ui.*;
+import com.badlogic.gdx.scenes.scene2d.ui.CheckBox;
+import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
-import com.badlogic.gdx.utils.Align;
 
-public class RegularEmitterDisplay extends Table {
+public class RegularEmitterDisplay extends ControllerComponentDisplay {
+
+	private final RegularEmitter emitter;
 
 	public RegularEmitterDisplay(RegularEmitter emitter, Skin skin) {
-		super(skin);
-		setBackground(skin.getDrawable("panel_thin_default"));
-		defaults().growX().minHeight(26f).space(4f);
-		Label header = new Label("Regular Emitter", skin);
-		header.setAlignment(Align.center);
-		add(header).row();
-		CheckBox continuousCheckbox = new CheckBox("Continuous", skin);
+		super("Regular Emitter", skin);
+		this.emitter = emitter;
+	}
+
+	@Override
+	void onUnfolded() {
+		CheckBox continuousCheckbox = new CheckBox("Continuous", getSkin());
 		continuousCheckbox.setChecked(emitter.isContinuous());
 		continuousCheckbox.addListener(new ChangeListener() {
 
@@ -24,13 +26,13 @@ public class RegularEmitterDisplay extends Table {
 				emitter.setContinuous(continuousCheckbox.isChecked());
 			}
 		});
-		add(continuousCheckbox).row();
-		add(new DisplayComponent.CountComponent(emitter, skin)).row();
-		add(new DisplayComponent.RangedNumericComponent("Delay", emitter.delayValue, skin)).row();
-		add(new DisplayComponent.RangedNumericComponent("Duration", emitter.durationValue, skin)).row();
-		add(new DisplayComponent.ScaledNumericComponent("Emission", emitter.emissionValue, skin)).row();
-		add(new DisplayComponent.ScaledNumericComponent("Life", emitter.lifeValue, skin)).row();
-		add(new DisplayComponent.ScaledNumericComponent("Life Offset", emitter.lifeOffsetValue, skin)).row();
+		addContent(continuousCheckbox);
+		addContent(new DisplayComponent.CountComponent(emitter, getSkin()));
+		addContent(new DisplayComponent.RangedNumericComponent("Delay", emitter.delayValue, getSkin()));
+		addContent(new DisplayComponent.RangedNumericComponent("Duration", emitter.durationValue, getSkin()));
+		addContent(new DisplayComponent.ScaledNumericComponent("Emission", emitter.emissionValue, getSkin()));
+		addContent(new DisplayComponent.ScaledNumericComponent("Life", emitter.lifeValue, getSkin()));
+		addContent(new DisplayComponent.ScaledNumericComponent("Life Offset", emitter.lifeOffsetValue, getSkin()));
 	}
 
 }